Update the answer action.
The answer action now refers to the same answering logic as the package mdr_question_answering. It uses the same source for its answers and is adjusted to the new syntax in the answers.txt file. Additionally, the listen action matches the input sentence to a registered sentence in either questions.txt or commands.txt in the mdr_speech_matching package. It therefore uses the SpeechMatching class from the same package.