Add further details for store grocery state machine.
This commit adds the states listen and process_command to the state machine Also the package.xml and CMakeList have been adjust accordingly. What needs to be done in an upcoming commit is: Create launch files for the actions listen and execute_command and maintain the launch file of the store_groceries launch file.