Add description and best case of the scenario.
This commit provides a description of the task, the sub-tasks, the goal and restrictions. Addtionally, a first state machine of the scenario is defined. The state machine focuses on one loop only and covers the best case of the scenario. In following commits further adjustments to the state machine need to be done.