[push_action] Refactored the action
I've particularly parameterised the action just as all other actions.
The action goal definition has been changed as well; the goal now
includes the pose of the object to be pushed, the pose of the goal
region where the object needs to be pushed, and the goal tolerance (the
tolerance is what defines the goal as a region rather than as just a
pose). This means that pushing can now be done along both x and y, not
just along x.

Unfortunately, the refactoring meant that I had to (hopefully
only temporarily) remove the fault detection part because, based on the
results, this didn't seem to be very reliable for practical purposes.