[wrs_tidy_up/config] Added a YAML file mapping from object classes to storage locations
For this particular case, the file acts as a "pseudo-ontology" (this is
essentially a model of the defaultStoringLocation relation in our
ontology).

We don't use the ontology for the scenario since the model
requires a reasoner, which is only supported by owlready2, which in turn
doesn't support Python 2 anymore (and we are still stuck with Python 2).