[move_arm_action/dmp] Fixed the online execution
The existing implementation was unrolling the DMP (i.e. calling the
dmp.step function) at every step of the control loop, without checking
if the intermediate goal has been reached. For this reason, the robot
was very quickly just starting to move directly towards the goal, thus
not following the trajectory at all.

In the fixed version, a DMP step is only taken if the currently pursued
goal has been reached, which means that the trajectory is actually
followed.