[pickup_action/action_states] Change grasping workflow to account for more errors
The newly implemented strategy uses neither MoveIt! nor DMPs; instead, a
simpler controller is implemented that aligns the base with the goal,
moves the base towards it, and, in the case of a top-down grasp, moves
the arm down for the object to be grasped. Care is taken to realign the
base with the goal at multiple steps in the workflow; otherwise, an
object is missed too often because of base misalignment